requirements.txt
# GitOps 2.0 Healthcare Intelligence - Python Dependencies
# Updated: December 2025 - Latest secure versions

# AI/ML Dependencies
openai>=1.59.0  # OpenAI API client - latest stable version

# Configuration & Data Processing
pyyaml>=6.0.2   # YAML parser - fixes CVE-2020-14343
requests>=2.32.3  # HTTP library - fixes multiple CVEs

# CLI & UX Dependencies
click>=8.1.0  # CLI framework for git_policy, ai_readiness, git_forensics
rich>=13.9.0  # Rich terminal formatting and progress bars

# Git Integration
gitpython>=3.1.0  # Git repository interaction for forensics

# Security & Compliance
cryptography>=44.0.0  # Cryptographic recipes - latest security fixes
pydantic>=2.10.0  # Data validation for config files

# Azure Cloud Services
azure-cosmos>=4.7.0  # Azure Cosmos DB SDK - NoSQL database
azure-identity>=1.19.0  # Azure authentication - managed identity support

# Testing & Development
pytest>=8.3.4  # Testing framework
pytest-asyncio>=0.24.0  # Async test support for Cosmos DB tests
pytest-cov>=6.0.0  # Coverage plugin
black>=24.12.0  # Code formatter
